+++ ci-internal/JenkinsfilePullRequestJob.groovy
@@ -83,6 +83,8 @@ def getTestInstrumentationRunnerName = { script, prefix ->
     ).split("\n").last()
 }
 
+def assembleCommandFileName = "buildSrc/build/tmp/assembleCommand.txt"
+
 //init
 def script = this
 def pipeline = new EmptyScmPipeline(script)
@@ -188,14 +190,14 @@ pipeline.stages = [
             script.sh "git merge origin/$destinationBranch --no-ff"
         },
 
-        pipeline.stage(RELEASE_NOTES_DIFF, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(RELEASE_NOTES_DIFF, StageStrategy.SKIP_STAGE) {
             script.sh("./gradlew generateReleaseNotesDiff -PrevisionToCompare=${lastDestinationBranchCommitHash}")
         },
 
-        pipeline.stage(CHECK_CONFIGURATION_IS_NOT_PROJECT_SNAPSHOT) {
+        pipeline.stage(CHECK_CONFIGURATION_IS_NOT_PROJECT_SNAPSHOT, StageStrategy.SKIP_STAGE) {
             script.sh "./gradlew checkConfigurationIsNotProjectSnapshotTask"
         },
-        pipeline.stage(CHECK_STABLE_MODULES_IN_ARTIFACTORY, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(CHECK_STABLE_MODULES_IN_ARTIFACTORY, StageStrategy.SKIP_STAGE) {
             withArtifactoryCredentials(script) {
                 script.echo "artifactory user: ${script.env.surf_maven_username}"
                 script.sh("./gradlew checkStableArtifactsExistInArtifactoryTask")
@@ -216,11 +218,11 @@ pipeline.stages = [
         pipeline.stage(CHECK_MODULES_IN_DEPENDENCY_TREE_OF_STABLE_MODULE_ALSO_STABLE, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
             script.sh("./gradlew checkStableComponentStandardDependenciesStableTask")
         },
-        pipeline.stage(CHECK_RELEASE_NOTES_VALID, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(CHECK_RELEASE_NOTES_VALID, StageStrategy.SKIP_STAGE) {
             script.sh("./gradlew checkReleaseNotesContainCurrentVersion")
             script.sh("./gradlew checkReleaseNotesNotContainCyrillic")
         },
-        pipeline.stage(CHECK_RELEASE_NOTES_CHANGED, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(CHECK_RELEASE_NOTES_CHANGED, StageStrategy.SKIP_STAGE) {
             script.sh("./gradlew checkReleaseNotesChanged -PrevisionToCompare=${lastDestinationBranchCommitHash}")
         },
         pipeline.stage(CHECKS_RESULT) {
@@ -248,9 +250,14 @@ pipeline.stages = [
         },
 
         pipeline.stage(BUILD) {
-            AndroidPipelineHelper.buildStageBodyAndroid(script, "clean assemble")
+            script.sh("./gradlew createCommandToAssembleOnlyChangedComponentsTask -PrevisionToCompare=${lastDestinationBranchCommitHash}")
+            String assembleCommand = script.readFile(assembleCommandFileName)
+
+            script.echo "Assemble components: $assembleCommand"
+
+            AndroidPipelineHelper.buildStageBodyAndroid(script, "clean $assembleCommand")
         },
-        pipeline.stage(BUILD_TEMPLATE, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(BUILD_TEMPLATE, StageStrategy.SKIP_STAGE) {
             script.sh("echo \"androidStandardDebugDir=$workspace\n" +
                     "androidStandardDebugMode=true\n" +
                     "skipSamplesBuild=true\" > template/android-standard/androidStandard.properties")
@@ -260,13 +267,13 @@ pipeline.stages = [
              */
             script.sh("./gradlew -p template clean build assembleQa assembleDebug --stacktrace")
         },
-        pipeline.stage(UNIT_TEST) {
+        pipeline.stage(UNIT_TEST, StageStrategy.SKIP_STAGE) {
             AndroidPipelineHelper.unitTestStageBodyAndroid(script,
                     "testReleaseUnitTest",
                     "**/test-results/testReleaseUnitTest/*.xml",
                     "app/build/reports/tests/testReleaseUnitTest/")
         },
-        pipeline.stage(INSTRUMENTATION_TEST_TEMPLATE, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(INSTRUMENTATION_TEST_TEMPLATE, StageStrategy.SKIP_STAGE) {
             script.dir("template") {
                 AndroidPipelineHelper.instrumentationTestStageBodyAndroid(
                         script,
@@ -283,7 +290,7 @@ pipeline.stages = [
                 )
             }
         },
-        pipeline.stage(INSTRUMENTATION_TEST, StageStrategy.UNSTABLE_WHEN_STAGE_ERROR) {
+        pipeline.stage(INSTRUMENTATION_TEST, StageStrategy.SKIP_STAGE) {
             AndroidPipelineHelper.instrumentationTestStageBodyAndroid(
                     script,
                     new AvdConfig(),