+++ easyadapter/lib-easyadapter/src/main/java/ru/surfstudio/android/easyadapter/ClassForDependentSample.kt
@@ -1,0 +1,4 @@
+package ru.surfstudio.android.easyadapter
+
+class ClassForDependentSample {
+}