+++ easyadapter/RELEASE_NOTES.md
@@ -16,6 +16,7 @@
   getItemHash to Object for BaseItemController; getItemId became
   abstract
 * ANDDEP-1017 Added DiffResultListener invoked after the DiffResult calculation
+<<<<<<< easyadapter/RELEASE_NOTES.md
 * ANDDEP-1038 **NO BACKWARD COMPATIBILITY**
   * Move `BasePaginationableAdapter` to easyadapter-pagination module
     and rename to `EasyPaginationAdapter`
@@ -31,8 +32,10 @@
   * Move `AnimatableViewHolder` and `BaseItemAnimator` to
     `recycler-extension` module
   * Rename `BaseItemAnimator` to `ViewHolderItemAnimator`
+* ANDDEP-1052 Added class for testing
 ##### Easyadapter Pagination
 * ANDDEP-1038 Add easyadapter-pagination module
+>>>>>>> easyadapter/RELEASE_NOTES.md
 ## 1.0.0
 ##### Easyadapter
 * ANDDEP-270 Added support for asynchronous view inflate in ViewHolder